requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
fastapi
sqlalchemy
pytest
pytest-asyncio
flake8==4.0.1
flake8-bugbear==22.1.11
flake8-pie==0.15.0
black
isort
uvicorn
aiosqlite
httpx
greenlet